+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/MostRecentSceneResponse.java
@@ -1,0 +1,13 @@
+package pl.cyfronet.s4e.controller.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.ZonedDateTime;
+
+@Data
+@Builder
+public class MostRecentSceneResponse {
+    private Long sceneId;
+    private ZonedDateTime timestamp;
+}